requirements/base.txt
boto3
coreapi
django
django-autocomplete-light
django-filter
django-pyodbc-azure
django-storages
django-taggit
django-taggit-serializer
djangorestframework
markdown
pillow
psycopg2-binary
python-dateutil
requests
sentry-sdk
uwsgi